Improve tests for `BranchClient.Search`
